STYLE: Rename header files to have the itk prefix.
Also fix trailing whitespace, indentation, some comments.
